Create TokenBase.sol
This is the open zeppelin erc20 contract used to create the LotSwap token. We added the mint/burn functions so that when we launch the mainnet bridge, cross chain transactions could occur. We have thought of possibly using a synthetic version of LotSwap for cross chain interaction and may decide not to use these functions. For safety we are also discussing updating the Admin to a multi sig wallet. If we decide on a synthetic version for mint/burn purposes we will redeploy the token to remove these functions.